python/pyproject.toml
[tool.poetry]
name = "artisan-builder"
version = "0.1.0"
description = "A build system for explainable science"
license = "MIT"
authors = ["Mason McGill <mmcgill@caltech.edu>"]
readme = "readme.rst"
packages = [
    { include = "artisan" }
]

[tool.poetry.dependencies]
dataclasses = "*"
cbor2 = "*"
falcon = "*"
gunicorn = "*"
h5py = "*"
fastjsonschema = "*"
numpy = "*"
"ruamel.yaml" = "*"
typing_extensions = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
